+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -2,6 +2,7 @@
 **2.1.2**
 - Exposed GameExtensions.GetResourceCacheCollection<T>(toolbox)
 - Added Lib.Gameplay.HasOpenSlotInInventory()
+- Added Vector3JsonConverter in the SOD.Common.Custom namespace
 
 **2.1.1**
 - Support for new workshop update