+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Exposed GameExtensions.GetResourceCacheCollection<T>(toolbox)
 - Added Lib.Gameplay.HasOpenSlotInInventory()
 - Added Vector3JsonConverter in the SOD.Common.Custom namespace
+- Added TimeDataJsonConverter in the SOD.Common.Custom namespace
 
 **2.1.1**
 - Support for new workshop update